Workable Version of Program
This version of the program works. It still needs adjustments to the "q" key event in order to make the program run faster, and it also needs to be adjusted in order to keep track of when both the fixation cross and the words appear and disappear on screen using a global timer. Program must also be adjusted to randomize 8 personal words 3 times, instead of 12 personal words 2 times, which it does currently.