Changed Quit Button from 'q' to 'escape'
The current MRI setup has a situation where pressing all 5 buttons on the button box simultaneously triggers a 'q' signal, which is the same signal that the previous version of this file recognized as the signal to terminate the program. This is bad because it effectively gives the participant the ability to terminate the program incredibly easily (either purposefully or accidentally) by pressing all 5 buttons at the same time while scanning.

The quit button signal has been changed from the 'q' button to the 'escape' button in order to avoid this issue.