Made Loops Selectable on Program Start
When starting the program, the technician can now choose whether to start from the Beginning of the task (Loop 1), Loop 2, or Loop 3. When a Loop is chosen, the program starts from that Loop and will perform all subsequent Loops afterwards. This is so that if an error occurs with the program while in the MRI room and a Loop was already completed, the program can be restarted without having to go through already completed Loops. Future Change Needed: Get Ready Message responds to t input when it should depend on time only.